Extract "to_basename" to utils.
