+++ pylib-v1/ailets/cons/util.py
@@ -1,0 +1,12 @@
+def to_basename(name: str) -> str:
+    """Return the base name of a node, stripping off any numeric suffix.
+
+    Args:
+        name: The full name of the node
+
+    Returns:
+        The base name of the node without the numeric suffix
+    """
+    if "." in name and name.split(".")[-1].isdigit():
+        return ".".join(name.split(".")[:-1])
+    return name
